Initial idea of custom dumpers with default to yaml_dumper.
This is just the initial idea that a 'dumper' argument can be passed
into `dump()` or `dumps()`. There are 2+1 dumpers provided:
`yaml_dumper()`, `toml_dumper()`, and `json_dumper()` which is empty now
in this commit.

And in order to make things even simpler, the delimiter is also embedded
into each dumper. That is, a dumper is responsibile for outputting the
delimiters.